added embedded H2 database, changed data type in Repo to match with model

+++ backend/src/main/java/com/project/craft/demo/resource/CityResource.java
@@ -30,7 +30,7 @@ public class CityResource {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<City> getCityDescription(@PathVariable("id") Long id) {
+    public ResponseEntity<City> getCityDescription(@PathVariable("id") Integer id) {
         City city = cityService.getCityDescriptionById(id);
         return new ResponseEntity<>(city, HttpStatus.OK);
     }
